+++ app/src/main/java/ru/practicum/android/diploma/di/DatabaseModule.kt
@@ -4,7 +4,7 @@ import androidx.room.Room
 import com.google.gson.Gson
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
-import ru.practicum.android.diploma.db.data.AppDatabase
+import ru.practicum.android.diploma.favourites.data.db.AppDatabase
 
 val dataBaseModule = module {
     single {
