+++ app/src/main/java/ru/practicum/android/diploma/di/InteractorModule.kt
@@ -1,8 +1,8 @@
 package ru.practicum.android.diploma.di
 
 import org.koin.dsl.module
-import ru.practicum.android.diploma.db.domain.interactor.FavouriteVacancyInteractor
-import ru.practicum.android.diploma.db.domain.interactor.FavouriteVacancyInteractorImpl
+import ru.practicum.android.diploma.favourites.domain.db.FavouriteVacancyInteractor
+import ru.practicum.android.diploma.favourites.domain.FavouriteVacancyInteractorImpl
 import ru.practicum.android.diploma.filter.domain.IndustriesInteractor
 import ru.practicum.android.diploma.filter.domain.IndustriesInteractorImpl
 import ru.practicum.android.diploma.filter.domain.SharedPrefInteractor
