+++ app/src/main/java/ru/practicum/android/diploma/favourites/data/db/AppDatabase.kt
@@ -1,9 +1,9 @@
-package ru.practicum.android.diploma.db.data
+package ru.practicum.android.diploma.favourites.data.db
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import ru.practicum.android.diploma.db.data.dao.VacancyDao
-import ru.practicum.android.diploma.db.data.entity.VacancyEntity
+import ru.practicum.android.diploma.favourites.data.db.entity.VacancyEntity
+import ru.practicum.android.diploma.favourites.data.db.dao.VacancyDao
 
 @Database(
     version = 1,
