+++ app/src/main/java/ru/practicum/android/diploma/favourites/data/db/dao/VacancyDao.kt
@@ -1,10 +1,10 @@
-package ru.practicum.android.diploma.db.data.dao
+package ru.practicum.android.diploma.favourites.data.db.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import ru.practicum.android.diploma.db.data.entity.VacancyEntity
+import ru.practicum.android.diploma.favourites.data.db.entity.VacancyEntity
 
 @Dao
 interface VacancyDao {
