--- app/src/main/java/ru/practicum/android/diploma/search/data/.gitkeep
@@ -1,1 +1,0 @@
-
