+++ app/src/main/java/ru/practicum/android/diploma/vacancy/domain/VacancyInteractor.kt
@@ -1,6 +1,6 @@
 package ru.practicum.android.diploma.vacancy.domain
 
-import ru.practicum.android.diploma.db.domain.interactor.FavouriteVacancyInteractor
+import ru.practicum.android.diploma.favourites.domain.db.FavouriteVacancyInteractor
 import ru.practicum.android.diploma.network.domain.VacancyNetworkRepository
 import ru.practicum.android.diploma.network.domain.models.ApiResult
 import ru.practicum.android.diploma.network.domain.models.VacancyDetail
